+++ CMake/GitExternal.cmake
@@ -14,13 +14,12 @@
 #    or in the given absolute path using the given repository and tag
 #    (commit-ish).
 #
+# Targets:
+#  * <directory>-rebase: fetches latest updates and rebases the given external
+#    git repository onto it.
+#  * rebase: Rebases all git externals, including sub projects
+#
 # Options which control behaviour:
-#  GIT_EXTERNAL_DISABLE_UPDATE
-#    When set, GitExternal will not change a repo that has already
-#    been checked out. The purpose of this is to allow one to set a
-#    default branch to be checked out, but stop GitExternal from
-#    changing back to that branch if the user has checked out and is
-#    working on another.
 #  GIT_EXTERNAL_VERBOSE
 #    When set, displays information about git commands that are executed
 #
@@ -37,7 +36,6 @@ if(NOT GIT_EXECUTABLE)
 endif()
 
 include(CMakeParseArguments)
-option(GIT_EXTERNAL_DISABLE_UPDATE "Disable update of cloned repositories" OFF)
 option(GIT_EXTERNAL_VERBOSE "Print git commands as they are executed" OFF)
 
 set(GIT_EXTERNAL_USER $ENV{GIT_EXTERNAL_USER})
@@ -81,6 +79,7 @@ function(GIT_EXTERNAL DIR REPO TAG)
   get_filename_component(GIT_EXTERNAL_DIR "${DIR}/.." ABSOLUTE)
 
   if(NOT EXISTS "${DIR}")
+    # clone
     message(STATUS "git clone ${REPO} ${DIR}")
     execute_process(
       COMMAND "${GIT_EXECUTABLE}" clone --recursive "${REPO}" "${DIR}"
@@ -88,6 +87,15 @@ function(GIT_EXTERNAL DIR REPO TAG)
       WORKING_DIRECTORY "${GIT_EXTERNAL_DIR}")
     if(nok)
       message(FATAL_ERROR "${DIR} git clone failed: ${error}\n")
+    endif()
+
+    # checkout requested tag
+    execute_process(
+      COMMAND "${GIT_EXECUTABLE}" checkout -q "${TAG}"
+      RESULT_VARIABLE nok ERROR_VARIABLE error
+      WORKING_DIRECTORY "${DIR}")
+    if(nok)
+      message(FATAL_ERROR "git checkout ${TAG} in ${DIR} failed: ${error}\n")
     endif()
   endif()
 
@@ -101,67 +109,76 @@ function(GIT_EXTERNAL DIR REPO TAG)
       WORKING_DIRECTORY "${DIR}")
   endif()
 
-  if(NOT IS_DIRECTORY "${DIR}/.git")
-    message(STATUS "Can't update git external ${DIR}: Not a git repository")
+  file(RELATIVE_PATH __dir ${CMAKE_SOURCE_DIR} ${DIR})
+  string(REGEX REPLACE "[:/]" "-" __target "${__dir}")
+  if(TARGET ${__target}-rebase)
     return()
   endif()
 
-  if(GIT_EXTERNAL_DISABLE_UPDATE)
-    git_external_message("git update disabled by user")
-    return()
-  endif()
+  set(__rebase_cmake "${CMAKE_CURRENT_BINARY_DIR}/${__target}-rebase.cmake")
+  file(WRITE ${__rebase_cmake}
+    "if(NOT IS_DIRECTORY ${DIR}/.git)\n"
+    "  message(FATAL_ERROR \"Can't update git external ${__dir}: Not a git repository\")\n"
+    "endif()\n"
+    # check if we are already on the requested tag
+    "execute_process(COMMAND ${GIT_EXECUTABLE} rev-parse --short HEAD\n"
+    "  OUTPUT_VARIABLE currentref OUTPUT_STRIP_TRAILING_WHITESPACE\n"
+    "  WORKING_DIRECTORY ${DIR})\n"
+    "if(currentref STREQUAL ${TAG}) # nothing to do\n"
+    "  return()\n"
+    "endif()\n"
+    "\n"
+    # reset generated files
+    "foreach(GIT_EXTERNAL_RESET_FILE ${GIT_EXTERNAL_RESET})\n"
+    "  execute_process(\n"
+    "    COMMAND \"${GIT_EXECUTABLE}\" reset -q \"\${GIT_EXTERNAL_RESET_FILE}\"\n"
+    "    ERROR_QUIET OUTPUT_QUIET\n"
+    "    WORKING_DIRECTORY \"${DIR}\")\n"
+    "  execute_process(\n"
+    "    COMMAND \"${GIT_EXECUTABLE}\" checkout -q -- \"${GIT_EXTERNAL_RESET_FILE}\"\n"
+    "    ERROR_QUIET OUTPUT_QUIET\n"
+    "    WORKING_DIRECTORY \"${DIR}\")\n"
+    "endforeach()\n"
+    "\n"
+    # fetch latest update
+    "execute_process(COMMAND \"${GIT_EXECUTABLE}\" fetch origin -q\n"
+    "  RESULT_VARIABLE nok ERROR_VARIABLE error\n"
+    "  WORKING_DIRECTORY \"${DIR}\")\n"
+    "if(nok)\n"
+    "  message(FATAL_ERROR \"Fetch for ${__dir} failed:\n   \${error}\")\n"
+    "endif()\n"
+    "\n"
+    # update tag
+    "execute_process(COMMAND ${GIT_EXECUTABLE} rebase FETCH_HEAD\n"
+    "  RESULT_VARIABLE nok ERROR_VARIABLE error OUTPUT_VARIABLE output\n"
+    "  WORKING_DIRECTORY \"${DIR}\")\n"
+    "if(nok)\n"
+    "  execute_process(COMMAND ${GIT_EXECUTABLE} rebase --abort\n"
+    "    WORKING_DIRECTORY \"${DIR}\" ERROR_QUIET OUTPUT_QUIET)\n"
+    "  message(FATAL_ERROR \"Rebase ${__dir} failed:\n\${output}\${error}\")\n"
+    "endif()\n"
+    "\n"
+    # checkout requested tag
+    "execute_process(\n"
+    "  COMMAND \"${GIT_EXECUTABLE}\" checkout -q \"${TAG}\"\n"
+    "  RESULT_VARIABLE nok ERROR_VARIABLE error\n"
+    "  WORKING_DIRECTORY \"${DIR}\")\n"
+    "if(nok)\n"
+    "  message(FATAL_ERROR \"git checkout ${TAG} in ${__dir} failed: ${error}\n\")\n"
+    "endif()\n"
+    )
 
-  # update to given tag
-  execute_process(COMMAND ${GIT_EXECUTABLE} rev-parse --short HEAD
-    OUTPUT_VARIABLE currentref OUTPUT_STRIP_TRAILING_WHITESPACE
-    WORKING_DIRECTORY ${DIR})
-  git_external_message(
-    "current ref is \"${currentref}\" and tag is \"${TAG}\"")
-  if(currentref STREQUAL TAG) # nothing to do
-    return()
+  add_custom_target(${__target}-rebase
+    COMMAND ${CMAKE_COMMAND} -P ${__rebase_cmake}
+    COMMENT "Rebasing ${__dir}")
+  set_target_properties(${__target}-rebase PROPERTIES
+    EXCLUDE_FROM_DEFAULT_BUILD ON FOLDER "git")
+  if(NOT TARGET rebase)
+    add_custom_target(rebase)
+    set_target_properties(rebase PROPERTIES EXCLUDE_FROM_DEFAULT_BUILD ON
+      FOLDER "git")
   endif()
-
-  # reset generated files
-  foreach(GIT_EXTERNAL_RESET_FILE ${GIT_EXTERNAL_RESET})
-    git_external_message("git reset -q ${GIT_EXTERNAL_RESET_FILE}")
-    execute_process(
-      COMMAND "${GIT_EXECUTABLE}" reset -q "${GIT_EXTERNAL_RESET_FILE}"
-      RESULT_VARIABLE nok ERROR_VARIABLE error
-      WORKING_DIRECTORY "${DIR}")
-    git_external_message("git checkout -q -- ${GIT_EXTERNAL_RESET_FILE}")
-    execute_process(
-      COMMAND "${GIT_EXECUTABLE}" checkout -q -- "${GIT_EXTERNAL_RESET_FILE}"
-      RESULT_VARIABLE nok ERROR_VARIABLE error
-      WORKING_DIRECTORY "${DIR}")
-  endforeach()
-
-  # fetch latest update
-  execute_process(COMMAND "${GIT_EXECUTABLE}" fetch origin -q
-    RESULT_VARIABLE nok ERROR_VARIABLE error
-    WORKING_DIRECTORY "${DIR}")
-  if(nok)
-    message(STATUS "Update of ${DIR} failed:\n   ${error}")
-  endif()
-
-  # update tag
-  git_external_message("git rebase FETCH_HEAD")
-  execute_process(COMMAND ${GIT_EXECUTABLE} rebase FETCH_HEAD
-    RESULT_VARIABLE RESULT ERROR_QUIET OUTPUT_QUIET
-    WORKING_DIRECTORY "${DIR}")
-  if(RESULT)
-    message(STATUS "git rebase failed, aborting ${DIR} merge")
-    execute_process(COMMAND ${GIT_EXECUTABLE} rebase --abort
-      WORKING_DIRECTORY "${DIR}" ERROR_QUIET OUTPUT_QUIET)
-  endif()
-
-  # checkout requested tag
-  execute_process(
-    COMMAND "${GIT_EXECUTABLE}" checkout -q "${TAG}"
-    RESULT_VARIABLE nok ERROR_VARIABLE error
-    WORKING_DIRECTORY "${DIR}")
-  if(nok)
-    message(STATUS "git checkout ${TAG} in ${DIR} failed: ${error}\n")
-  endif()
+  add_dependencies(rebase ${__target}-rebase)
 endfunction()
 
 set(GIT_EXTERNALS ${GIT_EXTERNALS_FILE})