+++ eq/fabric/configParams.h
@@ -73,6 +73,12 @@ public:
 
     /** @name Data Access. */
     //@{
+    /** Set the name of the configuration. @version 1.10 */
+    EQFABRIC_API void setName( const std::string& name );
+
+    /** @return the name of the configuration. @version 1.10 */
+    EQFABRIC_API const std::string& getName() const;
+
     /**
      * Set the name of the render client executable.
      *