+++ eq/qt/pipe.cpp
@@ -1,6 +1,5 @@
 
-/* Copyright (c) 2011, Stefan Eilemann <eile@eyescale.h>
- *               2012, Daniel Nachbaur <danielnachbaur@gmail.com>
+/* Copyright (c) 2015, Stefan.Eilemann@epfl.ch
  *
  * This library is free software; you can redistribute it and/or modify it under
  * the terms of the GNU Lesser General Public License version 2.1 as published
@@ -16,25 +15,31 @@
  * 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA.
  */
 
-#ifndef EQSERVER_CONFIG_DISPLAY_H
-#define EQSERVER_CONFIG_DISPLAY_H
+#include "pipe.h"
 
-#include "../types.h"
+#include <eq/pipe.h>
+#include <QDesktopWidget>
 
 namespace eq
 {
-namespace server
+namespace qt
 {
-namespace config
+bool Pipe::configInit()
 {
+    eq::Pipe* pipe = getPipe();
+    if( pipe->getPixelViewport().isValid( ))
+        return true;
 
-class Display
-{
-public:
-    static void discoverLocal( Config* config, const ConfigParams& params );
-};
+    QDesktopWidget desktop;
+    const uint32_t device = getPipe()->getDevice();
+    const int qtScreen = device == LB_UNDEFINED_UINT32 ?
+                             desktop.primaryScreen() : int( device );
+    const QRect rect = desktop.availableGeometry( qtScreen );
+
+    pipe->setPixelViewport( PixelViewport( rect.x(), rect.y(),
+                                           rect.width(), rect.height( )));
+    return true;
+}
 
 }
 }
-}
-#endif // EQSERVER_CONFIG_DISPLAY_H