+++ eq/util/accum.cpp
@@ -167,7 +167,6 @@ uint32_t Accum::getMaxSteps() const
 
     GLint accumBits;
     glGetIntegerv( GL_ACCUM_RED_BITS, &accumBits );
-
     return accumBits >= 16 ? 256 : 0;
 }
 
@@ -216,4 +215,3 @@ bool Accum::usesFBO( const GLEWContext* glewContext )
 
 }
 }
-