+++ examples/eqPly/initData.cpp
@@ -1,7 +1,7 @@
 
 /*
- * Copyright (c) 2006-2013, Stefan Eilemann <eile@equalizergraphics.com>
- *                    2010, Cedric Stalder <cedric.stalder@gmail.com>
+ * Copyright (c) 2006-2015, Stefan Eilemann <eile@equalizergraphics.com>
+ *                          Cedric Stalder <cedric.stalder@gmail.com>
  *
  * Redistribution and use in source and binary forms, with or without
  * modification, are permitted provided that the following conditions are met:
@@ -48,6 +48,10 @@ InitData::InitData()
     , _windowSystem( "GLX" )
 #elif WGL
     , _windowSystem( "WGL" )
+#elif EQ_QT_USED
+    , _windowSystem( "Qt" )
+#else
+#  error Unknown window system
 #endif
     , _renderMode( triply::RENDER_MODE_DISPLAY_LIST )
     , _useGLSL( false )